Data Collection and Preparation
In this code the data of the lowest round way flight ticket between Moscow (MOW) and St Petersburg is calculated. These are the lowest prices from the month of April 2020 to March 2021. It will vary if the API is called in the later months.The prices are given in Russian Rubel (RUB). This analysis is helpful for the passengers travelling between these two cities. It shows the lowest price for the months so that they can plan their journey accordingly.